x86: Store offset as int64_t and create a builder function for emission

+++ include/codegen/x86/MachineInst.hh
@@ -39,7 +39,7 @@ struct Operand {
     union {
         std::uint64_t imm;
         const void *lbl;
-        std::int8_t off;
+        std::int64_t off;
         Register reg;
     };
 };
